refactor async graph adapters
